Add create/delete file/directory
