+++ Menu.h
@@ -1,5 +1,9 @@
 #pragma
+#include <regex>
+#include <string>
+
 #include "FileManager.h"
+
 class Menu {
  private:
   FileManager fm;
@@ -8,6 +12,9 @@ class Menu {
 
   void IstepAhead();
   void IstepBack();
+  void IcreateFile();
+  void IcreateDirectory();
+  void IdeleteFileDirectory(const size_t index);
 
  public:
   void introPage();
