+++ Training.vcxproj.user
