--- Training/x64/Debug/Training.log
@@ -1,2 +1,0 @@
-﻿  FileManager.cpp
-  Training.vcxproj -> D:\LabShare\Training\x64\Debug\Training.exe
