--- Training/x64/Debug/Training.tlog/Training.lastbuildstate
@@ -1,2 +1,0 @@
-PlatformToolSet=v143:VCToolArchitecture=Native64Bit:VCToolsVersion=14.34.31933:TargetPlatformVersion=10.0.22621.0:
-Debug|x64|D:\LabShare\Training\|
